Code update, some bug fixes, UX improved
